requirements.txt
Django>=3.2.5
django-extensions>=3.1.3
djangorestframework>=3.12.4
APScheduler>=3.7.0
yahoo-fin>=0.8.8
yfinance>=0.1.64
drf-yasg>=1.17.1
pandas>=1.0.3
requests>=2.26.0
django_configurations>=2.3.1
dj_database_url>=0.5.0
django_debug_toolbar>=3.2.2